+++ src/powershell/Public/New-FinOpsCostExport.ps1
@@ -22,7 +22,7 @@
     Optional. Dataset to export. Allowed values = "ActualCost", "AmortizedCost", "FocusCost", "PriceSheet", "ReservationDetails", "ReservationRecommendations", "ReservationTransactions". Default = "FocusCost".
 
     .PARAMETER DatasetVersion
-    Optional. Schema version of the dataset to export. Default = "1.0" (applies to FocusCost only).
+    Optional. Schema version of the dataset to export. Default = "1.2-preview" (applies to FocusCost only).
 
     .PARAMETER DatasetFilters
     Optional. Dictionary of key/value pairs to filter the dataset with. Only applies to ReservationRecommendations dataset in 2023-07-01-preview. Valid filters are reservationScope (Shared or Single), resourceType (for example, VirtualMachines), lookBackPeriod (Last7Days, Last30Days, Last60Days).
@@ -330,8 +330,7 @@ function New-FinOpsCostExport
                 {
                     if ($Dataset -eq "FocusCost")
                     {
-                        # TODO: Change to "1.2-preview" after released
-                        $DatasetVersion = "1.0"
+                        $DatasetVersion = "1.2-preview"
                     }
                     elseif ($Dataset -eq "ActualCost" -or $Dataset -eq "AmortizedCost")
                     {